Bugfix: scaling factor was calculated using an integer division, not a double division.
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/trunk@225509 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/fop/layoutmgr/inline/ExternalGraphicLayoutManager.java
@@ -126,8 +126,8 @@ public class ExternalGraphicLayoutManager extends LeafNodeLayoutManager {
                     / fobj.getIntrinsicWidth());
             } else {
                 // adjust the larger
-                double rat1 = cwidth / fobj.getIntrinsicWidth();
-                double rat2 = cheight / fobj.getIntrinsicHeight();
+                double rat1 = (double)cwidth / fobj.getIntrinsicWidth();
+                double rat2 = (double)cheight / fobj.getIntrinsicHeight();
                 if (rat1 < rat2) {
                     // reduce cheight
                     cheight = (int)(rat1 * fobj.getIntrinsicHeight());
